Move Sterror class from TsSharedMutex.h to its own include file. (#8710)
