+++ include/tscpp/util/Makefile.am
@@ -22,6 +22,7 @@ library_include_HEADERS = \
 	IntrusiveDList.h \
 	LocalBuffer.h \
 	PostScript.h \
+	Strerror.h \
 	string_view_util.h \
 	TextView.h \
 	TsSharedMutex.h
